# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PEI Backend API",
    "description": "Plant Engineering IDE Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects/{project_id}/diagrams": {
      "post": {
        "tags": [
          "diagrams"
        ],
        "summary": "Create Diagram",
        "operationId": "create_diagram_projects__project_id__diagrams_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagramCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagram"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "diagrams"
        ],
        "summary": "Read Project Diagrams",
        "operationId": "read_project_diagrams_projects__project_id__diagrams_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Diagram"
                  },
                  "title": "Response Read Project Diagrams Projects  Project Id  Diagrams Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/diagrams/{diagram_id}": {
      "get": {
        "tags": [
          "diagrams"
        ],
        "summary": "Read Diagram",
        "operationId": "read_diagram_diagrams__diagram_id__get",
        "parameters": [
          {
            "name": "diagram_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagram Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagram"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "diagrams"
        ],
        "summary": "Update Diagram",
        "operationId": "update_diagram_diagrams__diagram_id__put",
        "parameters": [
          {
            "name": "diagram_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagram Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagramUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagram"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/diagrams/{diagram_id}/export": {
      "get": {
        "tags": [
          "diagrams"
        ],
        "summary": "Export Diagram",
        "operationId": "export_diagram_diagrams__diagram_id__export_get",
        "parameters": [
          {
            "name": "diagram_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagram Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Diagram": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "diagram_type": {
            "type": "string",
            "title": "Diagram Type",
            "default": "PID"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "canonical_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Canonical Json"
          },
          "canonical_schema_version": {
            "type": "integer",
            "title": "Canonical Schema Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "project_id",
          "version",
          "status",
          "canonical_json",
          "canonical_schema_version",
          "created_at",
          "updated_at"
        ],
        "title": "Diagram"
      },
      "DiagramCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "diagram_type": {
            "type": "string",
            "title": "Diagram Type",
            "default": "PID"
          },
          "canonical_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Canonical Json"
          }
        },
        "type": "object",
        "required": [
          "name",
          "canonical_json"
        ],
        "title": "DiagramCreate"
      },
      "DiagramUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "canonical_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Json"
          }
        },
        "type": "object",
        "title": "DiagramUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}